// semFront/src/app/models/citaMedico-interface.ts
export interface CitaMedicoInterface{
    id?: string;
    fecha_cita?: string;
    hora_inicio?: string;
    hora_final?: string;
    tipo?: string;
    titulo?: string;
    costo?: string;
    notas?: string;
    telefono?: string;
    unidadEnvio?: string;
    estatus?: string;
    creador?: string;
    editor?: string;
    creacion?: string;
    ultimaActualizacion?: string;
    Paciente?: any;
    Medico?: string;
    count?: any;
    results?: any;    
}